remove render from Component
