+++ src/index.d.ts
@@ -117,11 +117,11 @@ export interface IComponent<State = any, Props = any, Vm = any> {
     nvReceiveProps?(nextProps: Props): void;
     render(): Promise<IComponent<State, Props, Vm>>;
     reRender(): Promise<IComponent<State, Props, Vm>>;
-    mountComponent(dom: Element): void;
-    componentsConstructor(dom: Element): void;
-    getPropsValue(valueList: any[], value: any): void;
-    buildProps(prop: any): any;
-    buildComponentScope(ComponentClass: any, props: any, dom: Element): IComponent<any, any, any>;
+    // mountComponent(dom: Element): void;
+    // componentsConstructor(dom: Element): void;
+    // getPropsValue(valueList: any[], value: any): void;
+    // buildProps(prop: any): any;
+    // buildComponentScope(ComponentClass: any, props: any, dom: Element): IComponent<any, any, any>;
 }
 
 export interface INvModule {
